👀 [Feat] 게시물 작성 화면 ViewModel
- 추후 Firebase 관련 코드 작성 필요
